Extract: Add/Remove field functionality into module
